Correccion de textos en home y muevo los screens a una nueva carpeta
