+++ App.js
@@ -3,10 +3,10 @@ import React, { useContext, useState } from 'react';
 import { NavigationContainer, DefaultTheme } from '@react-navigation/native';
 import { createBottomTabNavigator } from '@react-navigation/bottom-tabs';
 import { AntDesign } from "@expo/vector-icons";
-import Home from "./components/Home";
-import Search from "./components/Search";
-import Profile from "./components/Profile";
-import Logout from './components/Logout';
+import Home from "./components/screens/Home";
+import Search from "./components/screens/Search";
+import Profile from "./components/screens/Profile";
+import Logout from './components/screens/Logout';
 import Login from "./components/Login";
 import GlobalContext from "./components/global/context";
 import { navigationRef } from './utils/RootNavigation';
