+++ components/screens/Profile.js
@@ -1,9 +1,9 @@
 
 import React, { useContext, useState } from "react";
 import { StyleSheet, Text, TouchableOpacity, View } from 'react-native';
-import InfoProfile from './InfoProfile';
-import ScrollViewMovies from './ScrollViewMovies';
-import GlobalContext from './global/context';
+import InfoProfile from '../InfoProfile';
+import ScrollViewMovies from '../ScrollViewMovies';
+import GlobalContext from '../global/context';
